+++ src/FE/src/lib/formatDate.ts
@@ -1,4 +1,4 @@
-const DATE_FORMATTER_LONG = new Intl.DateTimeFormat('da-DK', { day: 'numeric', month: 'long', year: 'numeric' });
+const DATE_FORMATTER_LONG = new Intl.DateTimeFormat('da-DK', { day: 'numeric', month: 'short', year: 'numeric' });
 export function formatDateLong(value: string | null | undefined): string | null {
   if (!value) return null;
   const date = new Date(value);
